Rename files to keep a pattern
